Migrated the navbar code into a separate component and added snapshot tests
